make _montgomery less OO
will decide tomorrow if we want to revert this commit... probably

going from OO to this paradigm is easy; basically it's just s/self\.//g and
s/\(self, /(/ ... but going the other way is a lot of work. so, probably we
should revert this commit for now and discuss it with JJChiDguez before
going ahead further with this unorthodox style of python.