remove last globals
this creates faux-oo objects for the poly_mul and poly_redc modules to store
the F_p object they require. they're faux-oo instead of proper classes to avoid
needing to pollute the code with 'self.' everywhere; i still am not sure if i
like this this approach.

prior to this commit, a CSIDH object would stop working correctly after another
CSIDH object was instantiated with a different prime.

this change also might fix a bug in our previous half-refactored
poly_mul_init() code, in a codepath in qring_mul which we apparently aren't
reaching in our tests, where the 'negative_powers_of_two' symbol was previously
undefined but now is in-scope.