start using pkg_resources
the package can now be installed with "pip install ." or "python setup.py
install" without root (eg, in one's home directory or in a virtualenv).

this is not well-tested yet. i plan to refactor the file reading and writing
throughout the csidh module to be handled in a uniform fashion and make the
writing an explicit operation.